not sure what I did last, probably testing LCD screen. Going to rewrite the drvier using DMA now anyways

+++ Firmware/Drivers/TFT_LCD/inc/TFT_LCD_Q.h
@@ -14,7 +14,7 @@
 
 typedef struct _lcdData {
     uint16_t *Data;
-    uint16_t Command;
+    uint8_t Command;
     uint32_t Length;
     uint8_t font;
     uint16_t color;
